Eliminate redundant simulations
Each season simulation now begins on the first week that has not been fully played - previously, the season simulation checked every space in the schedule grid to see if it had been played, and simulated the game if an unplayed game was scheduled. Runtime has decreased slightly as a result.